deleted empty index.js file and renamed model to index.js
